wip: started refactoring of key result custom list
