+++ src/app/graphql/key-result/custom-list/index.ts
@@ -1,5 +1,3 @@
 import GraphQLKeyResultCustomListModule from './module'
 
-export { KeyResultCustomListObject } from './models'
-
 export default GraphQLKeyResultCustomListModule
