+++ src/app/graphql/key-result/custom-list/module.ts
@@ -1,16 +1,15 @@
 import { Module } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
 
-import { Railway } from 'src/app/providers'
+import GraphQLAuthzModule from 'src/app/graphql/authz'
 import appConfig from 'src/config/app'
 import DomainModule from 'src/domain/module'
 
 import GraphQLKeyResultCustomListResolver from './resolver'
-import GraphQLKeyResultCustomListService from './service'
 
 @Module({
-  imports: [ConfigModule.forFeature(appConfig), DomainModule],
-  providers: [GraphQLKeyResultCustomListResolver, GraphQLKeyResultCustomListService, Railway],
+  imports: [ConfigModule.forFeature(appConfig), DomainModule, GraphQLAuthzModule],
+  providers: [GraphQLKeyResultCustomListResolver],
 })
 class GraphQLKeyResultCustomListsModule {}
