+++ src/app/graphql/key-result/module.ts
@@ -2,15 +2,21 @@ import { Module } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
 
 import GraphQLAuthzModule from 'src/app/graphql/authz'
-import { Railway } from 'src/app/providers'
 import appConfig from 'src/config/app'
 import DomainModule from 'src/domain/module'
 
+import GraphQLKeyResultCustomListModule from './custom-list'
 import GraphQLKeyResultResolver from './resolver'
 
 @Module({
-  imports: [ConfigModule.forFeature(appConfig), DomainModule, GraphQLAuthzModule],
-  providers: [GraphQLKeyResultResolver, Railway],
+  imports: [
+    ConfigModule.forFeature(appConfig),
+    DomainModule,
+    GraphQLAuthzModule,
+    GraphQLKeyResultCustomListModule,
+  ],
+  providers: [GraphQLKeyResultResolver],
+  exports: [GraphQLKeyResultCustomListModule],
 })
 class GraphQLKeyResultsModule {}
