+++ src/domain/key-result/custom-list/constants.ts
