+++ src/domain/key-result/custom-list/dto.ts
@@ -6,10 +6,10 @@ import { KEY_RESULT_CUSTOM_LIST_BINDING } from './constants'
 export class KeyResultCustomListDTO {
   id: string
   user: UserDTO
-  title?: string
-  binding?: KEY_RESULT_CUSTOM_LIST_BINDING
-  rank?: Array<KeyResultDTO['id']>
   createdAt: Date
   updatedAt: Date
   userId: UserDTO['id']
+  title?: string
+  binding?: KEY_RESULT_CUSTOM_LIST_BINDING
+  rank?: Array<KeyResultDTO['id']>
 }
