+++ src/domain/key-result/custom-list/repository.ts
@@ -1,6 +1,6 @@
 import { EntityRepository, SelectQueryBuilder, WhereExpression } from 'typeorm'
 
-import DomainEntityRepository, { CONSTRAINT_TYPE } from 'src/domain/repository'
+import { CONSTRAINT_TYPE, DomainEntityRepository } from 'src/domain/entity'
 import { TeamDTO } from 'src/domain/team/dto'
 import { Team } from 'src/domain/team/entities'
 import { UserDTO } from 'src/domain/user/dto'
