+++ src/domain/key-result/module.ts
@@ -3,6 +3,7 @@ import { TypeOrmModule } from '@nestjs/typeorm'
 
 import databaseConfig from 'src/config/database/config'
 
+import DomainKeyResultCustomListModule from './custom-list'
 import DomainKeyResultRepository from './repository'
 import DomainKeyResultService from './service'
 
@@ -10,6 +11,7 @@ import DomainKeyResultService from './service'
   imports: [
     TypeOrmModule.forRoot(databaseConfig),
     TypeOrmModule.forFeature([DomainKeyResultRepository]),
+    DomainKeyResultCustomListModule,
   ],
   providers: [DomainKeyResultService],
   exports: [DomainKeyResultService],
