+++ src/domain/user/service.ts
@@ -7,6 +7,8 @@ import { User } from './entities'
 import DomainUserRepository from './repository'
 
 export interface DomainUserServiceInterface {
+  repository: DomainUserRepository
+
   buildUserFullName: (user: UserDTO) => string
   getUserFromSubjectWithTeamRelation: (authzSub: UserDTO['authzSub']) => Promise<User>
 }
